Rename procs package to scheduler
